refactor: abstract repeated code into the useScrollProgress hook
