+++ src/components/DataRoom/CryptoPunks/Content.tsx
@@ -1,9 +1,10 @@
 import { type RefObject } from 'react'
 import { Typography } from '@mui/material'
-import { motion, useScroll, useTransform } from 'framer-motion'
+import { motion, useTransform } from 'framer-motion'
 import LinksWrapper from '../LinksWrapper'
 import { useSafeDataRoomStats } from '@/hooks/useSafeDataRoomStats'
 import { useIsMediumScreen } from '@/hooks/useMaxWidth'
+import useScrollProgress from '@/hooks/useScrollProgress'
 import type { BaseBlock } from '@/components/Home/types'
 import css from './styles.module.css'
 
@@ -14,10 +15,7 @@ const Content = ({ containerRef, title, text, link }: BaseBlock & { containerRef
   const { cryptoPunksStoredPercentage } = useSafeDataRoomStats()
   const isMobile = useIsMediumScreen()
 
-  const { scrollYProgress } = useScroll({
-    target: containerRef,
-    offset: ['start end', 'end start'],
-  })
+  const { scrollYProgress } = useScrollProgress(containerRef)
 
   const opacityParams = isMobile ? [0.4, 0.45, 0.65, 0.66] : [0.25, 0.35, 0.65, 0.7]
   const opacity = useTransform(scrollYProgress, opacityParams, [0, 1, 1, 0])
