+++ src/components/DataRoom/CryptoPunks/PunksGrid.tsx
@@ -1,5 +1,6 @@
 import { type RefObject } from 'react'
-import { motion, useScroll, useTransform } from 'framer-motion'
+import { motion, useTransform } from 'framer-motion'
+import useScrollProgress from '@/hooks/useScrollProgress'
 import { getRandomColor } from '@/components/DataRoom/CryptoPunks/utils/getRandomColor'
 import CryptoPunk from '@/public/images/DataRoom/cryptopunk-silhouette.svg'
 import css from './styles.module.css'
@@ -8,10 +9,7 @@ const CRYPTOPUNK_ROWS_NR = 8
 const CRYPTOPUNK_COLUMNS_NR = 24
 
 const PunksGrid = ({ containerRef, isMobile }: { containerRef: RefObject<HTMLDivElement>; isMobile: boolean }) => {
-  const { scrollYProgress } = useScroll({
-    target: containerRef,
-    offset: ['start end', 'end start'],
-  })
+  const { scrollYProgress } = useScrollProgress(containerRef)
 
   const translateParams = isMobile ? [0, 1] : [0.25, 0.75]
   const opacity = useTransform(scrollYProgress, [0, 0.25, 0.7, 0.75], [0, 1, 1, 0])
