+++ src/hooks/useScrollProgress.ts
@@ -1,26 +1,10 @@
-import { useEffect, useState } from 'react'
+import type { RefObject } from 'react'
+import { useScroll } from 'framer-motion'
 
-const useScrollProgress = () => {
-  const [readProgress, setReadProgress] = useState(0)
-
-  const handleScroll = () => {
-    const windowHeight = window.innerHeight
-    const documentHeight = document.documentElement.scrollHeight
-    const scrollTop = window.scrollY || document.documentElement.scrollTop
-    const scrollPercent = (scrollTop / (documentHeight - windowHeight)) * 100
-
-    setReadProgress(scrollPercent)
-  }
-
-  useEffect(() => {
-    window.addEventListener('scroll', handleScroll)
-
-    return () => {
-      window.removeEventListener('scroll', handleScroll)
-    }
-  }, [])
-
-  return readProgress
-}
+const useScrollProgress = (containerRef: RefObject<HTMLElement>) =>
+  useScroll({
+    target: containerRef,
+    offset: ['start end', 'end start'],
+  })
 
 export default useScrollProgress
